+++ src/components/Picker/constants.js
@@ -1,0 +1,2 @@
+export const MODE_DIALOG = MODE_DIALOG;
+export const MODE_DROPDOWN = MODE_DROPDOWN;
